+++ lib/ncs_navigator/warehouse/data_mapper.rb
@@ -23,16 +23,36 @@ module NcsNavigator::Warehouse
       end
     end
 
+    module NcsStringType
+      protected
+
+      def self.included(into)
+        into.class_eval do
+          alias :original_typecast_to_primitive :typecast_to_primitive
+
+          def typecast_to_primitive(value)
+            if value.is_a?(BigDecimal)
+              value.to_s('F')
+            else
+              original_typecast_to_primitive(value)
+            end
+          end
+        end
+      end
+    end
+
     ##
     # DataMapper `:string` that's an {NcsType}.
     class NcsString < ::DataMapper::Property::String
       include NcsType
+      include NcsStringType
     end
 
     ##
     # DataMapper `:text` that's an {NcsType}.
     class NcsText < ::DataMapper::Property::Text
       include NcsType
+      include NcsStringType
     end
 
     ##