+++ lib/ncs_navigator/warehouse.rb
@@ -7,6 +7,7 @@ module NcsNavigator
     DEFAULT_MDES_VERSION = '2.0'
 
     autoload :CLI,                 'ncs_navigator/warehouse/cli'
+    autoload :Comparator,          'ncs_navigator/warehouse/comparator'
     autoload :Configuration,       'ncs_navigator/warehouse/configuration'
     autoload :DataMapper,          'ncs_navigator/warehouse/data_mapper'
     autoload :DatabaseInitializer, 'ncs_navigator/warehouse/database_initializer'